requirements.txt
# ============================================
# Desktop Web View
# ============================================
pywebview==5.3.1
pystray==0.19.5

# ============================================
# Packaging
# ============================================
pyinstaller==6.3.0

# ============================================
# Task Scheduling
# ============================================
APScheduler==3.10.4

# ============================================
# Secure Token Storage
# ============================================
keyring==25.5.0

# ============================================
# Automation & Mobile Testing (uiautomator2)
# ============================================
uiautomator2

# ============================================
# HTTP & API Communication
# ============================================
requests==2.32.3
httpx==0.28.1

# ============================================
# Data Processing & Validation
# ============================================
pydantic==2.10.5
python-dotenv==1.0.1

# ============================================
# Image Processing & Computer Vision
# ============================================
opencv-python==4.10.0.84
Pillow==11.1.0
numpy==2.2.2

# ============================================
# Utilities
# ============================================
pyperclip==1.9.0
python-dateutil==2.9.0.post0

# ============================================
# Development & Testing
# ============================================
pytest==8.3.4
black==24.10.0
flake8==7.1.1
